Dev to master merge

+++ src/main/java/com/microsoft/graph/requests/extensions/ChunkedUploadRequest.java
@@ -71,8 +71,8 @@ public class ChunkedUploadRequest {
                                 final byte[] chunk,
                                 final int chunkSize,
                                 final int maxRetry,
-                                final int beginIndex,
-                                final int totalLength) {
+                                final long beginIndex,
+                                final long totalLength) {
         this.data = new byte[chunkSize];
         System.arraycopy(chunk, 0, this.data, 0, chunkSize);
         this.retryCount = 0;
